requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.3.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
dataclass-factory==2.16
exceptiongroup==1.2.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.1.2
pillow==11.0.0
propcache==0.2.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
requests==2.32.3
shazamio==0.7.0
shazamio_core==1.0.7
sniffio==1.3.1
sounddevice==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
